+++ src/Outsystems.SetupTools/Functions/Install-OSServer.ps1
@@ -20,6 +20,9 @@ function Install-OSServer
     .PARAMETER Version
     The version to be installed.
 
+    .PARAMETER WithLifetime
+    If specified, the cmdlet will install the platform server with lifetime.
+
     .EXAMPLE
     Install-OSServer -Version "10.0.823.0"
 
@@ -44,17 +47,23 @@ function Install-OSServer
     param(
         [Parameter(ParameterSetName = 'Local')]
         [Parameter(ParameterSetName = 'Remote')]
+        [ValidateNotNullOrEmpty()]
         [string]$InstallDir = $OSDefaultInstallDir,
 
         [Parameter(ParameterSetName = 'Local', Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [string]$SourcePath,
 
         [Parameter(ParameterSetName = 'Local', Mandatory = $true)]
         [Parameter(ParameterSetName = 'Remote', Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
         [version]$Version,
 
         [Parameter()]
-        [switch]$SkipRabbitMQ
+        [switch]$SkipRabbitMQ,
+
+        [Parameter()]
+        [switch]$WithLifetime
     )
 
     begin
@@ -73,6 +82,14 @@ function Install-OSServer
 
         $osVersion = GetServerVersion
         $osInstallDir = GetServerInstallDir
+
+        # Installer name
+        $osInstaller = "PlatformServer-$Version.exe"
+        if ($WithLifetime.IsPresent)
+        {
+            # Lifetime installer instead
+            $osInstaller = "LifeTimeWithPlatformServer-$Version.exe"
+        }
     }
 
     process
@@ -167,12 +184,12 @@ function Install-OSServer
                 {
                     LogMessage -Function $($MyInvocation.Mycommand) -Phase 1 -Stream 0 -Message "SourcePath not specified. Downloading installer from repository"
 
-                    $installer = "$ENV:TEMP\PlatformServer-$Version.exe"
+                    $installer = "$ENV:TEMP\$osInstaller"
                     LogMessage -Function $($MyInvocation.Mycommand) -Phase 1 -Stream 0 -Message "Saving installer to $installer"
 
                     try
                     {
-                        DownloadOSSources -URL "$OSRepoURL\PlatformServer-$Version.exe" -SavePath $installer
+                        DownloadOSSources -URL "$OSRepoURL\$osInstaller" -SavePath $installer
                     }
                     catch
                     {
@@ -189,7 +206,7 @@ function Install-OSServer
                 "Local"
                 {
                     LogMessage -Function $($MyInvocation.Mycommand) -Phase 1 -Stream 0 -Message "SourcePath specified. Using the local installer"
-                    $installer = "$SourcePath\PlatformServer-$Version.exe"
+                    $installer = "$SourcePath\$osInstaller"
                 }
             }